requirements.txt
bleach==6.3.0
blinker==1.9.0
click==8.3.1
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-CKEditor==1.0.0
Flask-Limiter==4.1.1
Flask-Mail==0.10.0
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==25.0
python-dotenv==1.2.1
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.5
wrapt==2.0.1
WTForms==3.2.1
